New button element, login button, new mainpage, handleInput for button
Created a new element `button`, along with its rendering and input handling
Used this element to create a login button.

Each button takes a `action` which is a `proc`, so for the login button the action is `mainpageRoute()`,
which routes to the mainpage.

TODO:
To transfer control from login page to mainpage, I cant just import certain functions like the `initTUIState()` apparently I cant pass the same TUI state and have to initialize a new one, otherwise I'm violating memory safety is what the compiler screams. Need to look into this, but temporary solution was to re-write that procedure locally inside the `mainpageRoute()` proc and initilize a new TUIState.